pyproject.toml
[project]
name = "poke-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "fastmcp>=2.8.0",
    "google-genai>=1.20.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "rich>=13.0.0",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
